hcl: fix the test that pinned the old float behaviour
TestHCLToMapValueShapes asserted m["n"] == 3.0, which is the behaviour the
previous commit deliberately changed. It is a test of what the code did, not
of what it should do, and it had to move with the decision.

I pushed the previous commit with this test RED. `go test … | grep -v ok`
returns grep's status, not the suite's, so the chain that followed it ran as
though everything had passed. That is the same pipe-swallows-the-exit-status
mistake this project has recorded before, made while fixing something else.

+++ hcl_test.go
@@ -58,7 +58,10 @@ blk {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if m["s"] != "text" || m["b"] != true || m["n"] != 3.0 || m["f"] != 1.5 || m["nul"] != nil {
+	// n is an int64 and f a float64: an integral number keeps its integerness,
+	// because yaml.Marshal renders a large float64 in exponent form and a
+	// version pinned to 20250127 must not become 2.0250127e+07.
+	if m["s"] != "text" || m["b"] != true || m["n"] != int64(3) || m["f"] != 1.5 || m["nul"] != nil {
 		t.Errorf("scalars: %#v", m)
 	}
 	if l, ok := m["list"].([]any); !ok || len(l) != 2 {
